+++ postcss.config.js
@@ -1,8 +1,10 @@
 module.exports = {
   plugins: [
-      require('autoprefixer'),
-      require('cssnano')({ // подключили cssnano
-          preset: 'default', // выбрали настройки по умолчанию
-  })
-]
+    require('autoprefixer')({
+      browsers:'last 10 versions'
+    }),
+    require('cssnano')({ // подключили cssnano
+      preset: 'default', // выбрали настройки по умолчанию
+    })
+  ]
 }
