+++ src/index.html
@@ -75,13 +75,6 @@
       </div>
     </div>
   </div>
-  <!-- <script src="./scripts/Card.js"></script>
-  <script src="./scripts/CardList.js"></script>
-  <script src="./scripts/Popup.js"></script>
-  <script src="./scripts/UserInfo.js"></script>
-  <script src="./scripts/FormValidator.js"></script>
-  <script src="./scripts/Api.js"></script>
-  <script src="./scripts/script.js"></script> -->
   <script src="<%= htmlWebpackPlugin.files.js %>"></script>
 </body>
 </html>
